Enable pen, random, eraser buttons in UI
We renamed the original 'enableTool' functions to
'enableSquareTool' and created new enableTool
functions that call forEach methods on squares,
the match the button event listener callbacks.
Additionally, we decided to use inline styling
methods to color all of the squares to
maintain consistency and not overcomplicate
some of the functions.